requirements.txt
Keras==2.2.5
numpy==1.16.6
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==6.2.0
tensorflow-gpu==1.15.0
progressbar2
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI
tqdm==4.28.1